finished all pages, error handling not finished

+++ admin-crud/delete.php
@@ -7,6 +7,25 @@
     }
     if($_SERVER["REQUEST_METHOD"] == "GET"){
         $id = $mysqli->real_escape_string($_GET["id"]);
+
+        $sql_img = "SELECT * FROM products WHERE id = ?";
+        $img_stmt = $mysqli->stmt_init();
+        if(!$img_stmt->prepare($sql_img)){
+            die("SQL error: " . $mysqli->error);
+        }
+        $img_stmt->bind_param("i", $id);
+        $img_stmt->execute();
+        $img_result = $img_stmt->get_result();
+        if ($img_result->num_rows > 0) {
+            $img_row = $img_result->fetch_assoc();
+            $product_img = $img_row["product_img"];
+        }
+        $path = "../uploads/$product_img";
+        if(!unlink($path)){
+            header("Location: ../admin.php?error=img");
+            exit();
+        }
+
         $sql = "DELETE FROM products WHERE id=?;";
         $stmt = $mysqli->stmt_init();
         if(!$stmt->prepare($sql)){
@@ -14,7 +33,7 @@
         }
         $stmt->bind_param("i", $id);
         $stmt->execute();
-        header("Location: ../admin.php");
+        header("Location: ../admin.php?succes=delete");
         exit();
     } else {
         header("Location: ../admin.php");
